ai_architecture/requirements.txt
python-dotenv==1.0.1
PyPDF2==3.0.1
aiohttp==3.9.3
asyncio==3.4.3
python-pptx==0.6.23
python-docx==1.1.0
cachetools==5.3.2 